refactor go compiler
